% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opt_tmle.R
\name{assign_treatment}
\alias{assign_treatment}
\title{assign_treatment}
\usage{
assign_treatment(obj, newdata)
}
\arguments{
\item{obj,}{an object generated by opt_tmle}

\item{newdata,}{a data.frame containing columns corresponding to the Vnodes}
}
\description{
Given a opt_tmle object, assign treatment to a new dataset
}
\seealso{
\code{\link{opt_tmle}}
}

