% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opt_tmle.R
\docType{data}
\name{opt_tmle.SL.library}
\alias{opt_tmle.SL.library}
\title{Example library specification}
\format{\preformatted{List of 4
 $ Q    : chr [1:8] "SL.glm" "SL.glmem" "SL.glmnetprob" "SL.step.forward" ...
 $ g    : chr [1:4] "mnSL.randomForest" "mnSL.glmnet" "mnSL.multinom" "mnSL.mean"
 $ QaV  : chr [1:7] "SL.glm" "SL.glmnetprob" "SL.step.forward" "SL.gam" ...
 $ class: chr [1:4] "mnSL.randomForest" "mnSL.glmnet" "mnSL.multinom" "mnSL.mean"
}}
\usage{
opt_tmle.SL.library
}
\description{
Separate libraries must be specified for Q, g, QaV, and the classifier
}
\keyword{datasets}

