% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EM_SL_learners.R
\name{SL.glmnetprob}
\alias{SL.glmnetprob}
\title{glmnet wrapper that handles probabilities instead of just binary 1's and 0's}
\usage{
SL.glmnetprob(Y, X, newX, family, obsWeights, id, alpha = 1, nfolds = 10,
  nlambda = 100, useMin = TRUE, ...)
}
\description{
glmnet wrapper that handles probabilities instead of just binary 1's and 0's
}

