% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opt_tmle.R
\name{opt_tmle}
\alias{opt_tmle}
\alias{plot.opt_tmle}
\alias{predict.opt_tmle}
\alias{print.opt_tmle}
\title{opt_tmle}
\usage{
opt_tmle(data, Wnodes = grep("^W", names(data), value = TRUE), Anode = "A",
  Ynode = "Y", Vnodes = Wnodes, stratifyAY = TRUE,
  SL.library = opt_tmle.SL.library, verbose = 3, parallel = FALSE,
  perf_tmle = TRUE, perf_dripcw = FALSE, perf_cv = TRUE,
  perf_full = FALSE, maximize = TRUE, ...)

\method{print}{opt_tmle}(obj)

\method{plot}{opt_tmle}(obj)

\method{predict}{opt_tmle}(obj, ...)
}
\arguments{
\item{data}{data.frame containing the relevant variable}

\item{Wnodes,}{vector of column names indicating covariates}

\item{Anode,}{column name of treatment}

\item{Ynode,}{column name of outcome}

\item{Vnodes,}{vector of column names to base the treatment on}

\item{stratifyAY,}{logical: should we stratify the cross-validation based on (A,Y) pairs}

\item{verbose,}{integer that controls the verbosity of the output (higher is more verbose)}

\item{parallel,}{logical: should foreach parallelization be used?}

\item{SL_library,}{list of SuperLearner libraries for the various models. See \code{\link{opt_tmle.SL_library}} for an example.}
}
\description{
Estimation of the Optimal Treatment rule using Super Learner and mean performance using CV-TMLE
To avoid nesting cross-validation, it uses split-specfic estimates of Q and g to estimate the rule, and
'split-specific' estimates of the rule in CV-TMLE to estimate mean performance
}
\examples{
Qbar0 <- function(A, W) {
    
    W1 <- W[, 1]
    W2 <- W[, 2]
    W3 <- W[, 3]
    W4 <- W[, 4]
    Qbar <- plogis(ifelse(W4 > 0, (A == 1) + (A == 1) * (5 * W1^2 - 4.45), (A == 
        2) + (A == 3) + (A == 2) * (4 * W2) + (A == 3) * (5 * W3)))
    return(Qbar)
}

g0 <- function(W) {
    W1 <- W[, 1]
    W2 <- W[, 2]
    W3 <- W[, 3]
    W4 <- W[, 4]
    
    # rep(0.5, nrow(W))
    scale_factor <- 0.8
    A1 <- plogis(scale_factor * W1)
    A2 <- plogis(scale_factor * W2)
    A3 <- plogis(scale_factor * W3)
    A <- cbind(A1, A2, A3)
    
    # make sure A sums to 1
    A <- normalize_rows(A)
}

gen_data <- function(n = 1000, p = 4) {
    W <- matrix(rnorm(n * p), nrow = n)
    colnames(W) <- paste("W", seq_len(p), sep = "")
    pA <- g0(W)
    A <- factor(apply(pA, 1, function(pAi) which(rmultinom(1, 1, pAi) == 1)))
    A_vals <- vals_from_factor(A)
    
    u <- runif(n)
    Y <- as.numeric(u < Qbar0(A, W))
    Q0aW <- sapply(A_vals, Qbar0, W)
    d0 <- max.col(Q0aW)
    Yd0 <- as.numeric(u < Qbar0(d0, W))
    data.frame(W, A, Y, Q0aW, d0, Yd0)
}

testdata <- gen_data(1e+05, 5)


opt_tmle.SL.library <- list(Q = c("SL.glm", "SL.glmem", "SL.glmnetprob", "SL.step.forward", 
    "SL.gam", "SL.rpart", "SL.rpartPrune", "SL.mean"), g = c("mnSL.randomForest", 
    "mnSL.glmnet", "mnSL.multinom", "mnSL.mean"), QaV = c("SL.glm", "SL.glmnetprob", 
    "SL.step.forward", "SL.gam", "SL.rpart", "SL.rpartPrune", "SL.mean"), class = c("mnSL.randomForest", 
    "mnSL.glmnet", "mnSL.multinom", "mnSL.mean"))

data <- gen_data(1000, 5)
result <- opt_tmle(data, SL.library = opt_tmle.SL.library)
print(result)
Wnodes <- result$nodes$Wnodes
QaV_dV <- predict(result, newdata = testdata[, Wnodes], pred_fit = "QaV")
QaV_perf <- mean(Qbar0(QaV_dV, testdata[, Wnodes]))
class_dV <- predict(result, newdata = testdata[, Wnodes], pred_fit = "class")
class_perf <- mean(Qbar0(class_dV, testdata[, Wnodes]))
joint_dV <- predict(result, newdata = testdata[, Wnodes], pred_fit = "joint")
joint_perf <- mean(Qbar0(joint_dV, testdata[, Wnodes]))
EYd0_perf <- mean(Qbar0(testdata$d0, testdata[, Wnodes]))
c(QaV_perf, class_perf, joint_perf, EYd0_perf)
# perf of true blip approx=0.748

print(result)
plot(result)

vim <- tx_vim(result)
ggplot(vim, aes(y = node, x = risk_full_fraction, color = model)) + geom_point() + 
    theme_bw() + xlab("VIM")

library(reshape2)
long <- melt(vim, id = c("node", "model"))
ggplot(long, aes(y = node, x = value, color = model)) + geom_point() + facet_wrap(~variable, 
    scales = "free") + theme_bw() 
}

