% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/opttx-package.R
\docType{package}
\name{opttx-package}
\alias{opttx}
\alias{opttx-package}
\title{opttx - Optimal Treatment Estimation}
\description{
opttx - Optimal Treatment Estimation
}
\details{
\tabular{ll}{
Package: \tab origami\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2014-08-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr

}

opttx estimates optimal treatment rules and the mean outcome under such rules
}
\examples{
Qbar0 <- function(A, W) {
  W1 <- W[, 1]
  W2 <- W[, 2]
  W3 <- W[, 3]
  W4 <- W[, 4]
  Qbar <- ifelse(W4 > 0, plogis(1 - W1^2 + 3 * W2 + A * (5 * W3^2 - 4.45)), plogis(-0.5 - W3 + 2 * W1 * W2 + A * (3 * 
                                                                                                                    abs(W2) - 1.5)))
  return(Qbar)
}

g0 <- function(W) {
  rep(0.5, nrow(W))
}

gen_data <- function(n = 1000, p = 4) {
  W <- matrix(rnorm(n * p), nrow = n)
  colnames(W) <- paste("W", seq_len(p), sep = "")
  A <- rbinom(n, 1, g0(W))
  u <- runif(n)
  Y <- as.numeric(u < Qbar0(A, W))
  Y0 <- as.numeric(u < Qbar0(0, W))
  Y1 <- as.numeric(u < Qbar0(1, W))
  d0 <- as.numeric(Qbar0(1, W) > Qbar0(0, W))
  Yd0 <- as.numeric(u < Qbar0(d0, W))
  data.frame(W, A, Y, Y0, Y1, Yd0, d0, blip = Qbar0(1, W) - Qbar0(0, W))
}

data=gen_data(1000,5)
system.time({
  result=opt_tmle(data)
})

print(result)
plot(result)
}
\author{
Jeremy Coyle \email{jeremyrcoyle@gmail.com}
}
\seealso{
\code{\link{cross_validate}}, \code{\link{origami_SuperLearner}}
}
\keyword{package}

